tests/submission/MOSQUITO/input/range.html

<!DOCTYPE html>
<html>

  <head>
    <title>Input Range</title>
     <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
 	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
 	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
   <script src="/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  </head>

  <body>

  
  		<h1>Input Range</h1>
  		<input type="range" style="display:none" min=0 max=5 />

    <div id="log">
    </div>
	
	<script type="text/javascript">
		

		test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "range")}, "range type support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[0].min, "0")}, "min attribute support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[0].max, "5")}, "max attribute support on input element");
		  
	</script>
	
  </body>
  
</html>
